Control Panel: fix backup save/start failing on a non-existent COM method (#19)
Saving the backup settings ended with backup.Save(), but IInterfaceBackupSettings
has no Save method - every property setter writes straight through to the
server's settings store as it is assigned. The Control Panel reaches the server
over late-bound COM, so the bogus call compiled and only failed at runtime with
"'System.__ComObject' does not contain a definition for 'Save'". Because
"Start backup now" saves first and bails out when the save reports failure, the
backup never started; the settings themselves were already persisted by then.

This was the second instance of the class (the first was the Diagnostics page
reading Success/Details instead of Result/ExecutionDetails, fixed in 6.2.13), so
every late-bound COM member access in all 26 Control Panel files that use one was
audited against hMailServer.idl and the COM implementations. No further missing
or misspelled members were found. Two adjacent defects that audit surfaced are
fixed here as well:

  - QueueView narrowed the 64-bit message id to Int32 when retriggering
    delivery, so Retry threw OverflowException for ids above 2^31 on long-lived
    installations. Remove already used Int64.

  - The "email all accounts" utility discarded the server's boolean result and
    reported "Sendout queued" even when the server declined it.

Verified: Control Panel builds clean with -warnaserror; ControlPanel.Tests 17 of
17 passing. No server-core change.

+++ hmailserver/source/Tools/ControlPanel/Views/QueueView.xaml.cs
@@ -55,7 +55,9 @@ namespace hMailServer.ControlPanel.Views
          try
          {
             dynamic queue = ServerSession.Current.Application.GlobalObjects.DeliveryQueue;
-            queue.ResetDeliveryTime(Convert.ToInt32(row.Id));
+            // Message ids are 64-bit on the server; narrowing to Int32 threw
+            // OverflowException on long-lived installations (Remove already used Int64).
+            queue.ResetDeliveryTime(Convert.ToInt64(row.Id));
             queue.StartDelivery();
             ServerSession.Release(queue);
             SubtitleText.Text = "Delivery retriggered for message " + row.Id + ".";
